change uints to ints and remove extra metrics from analyzer

+++ sbndaq-artdaq/Generators/SBND/NevisTPC/NevisTB_generatorBase.hh
@@ -132,13 +132,13 @@ namespace sbndaq
     int current_ntbevent, current_ntbframenum; // for checking consistency between FEMs within a run
     bool desyncCrash;
     uint64_t pseudo_ntbfragment;
-    uint32_t rollCounter;
-    uint32_t prevFrame;
-    uint32_t prevEvent;
-    uint32_t prevCorrFrame;
-    uint8_t N_Bad_DEvents;
-    uint8_t deadtime_in_frames;
-    uint8_t N_Bad_DFrames;
+    int rollCounter;
+    int prevFrame;
+    int prevEvent;
+    int prevCorrFrame;
+    int N_Bad_DEvents;
+    int deadtime_in_frames;
+    int N_Bad_DFrames;
     long long GPSframe;
     long long GPSsample;
     long long GPSdiv;
